backend/requirements.txt
flask
youtube-transcript-api
flask-cors
requests
python-dotenv
googletrans==4.0.0rc1 